pyproject.toml
[project]
name = "otel-demo"
version = "0.0.1"
description = "Opentelemetry Demo"
authors = [
    {name = "Mikołaj Nowak", email = "12396461+m3nowak@users.noreply.github.com"},
]
dependencies = ["opentelemetry-sdk>=1.27.0", "grpcio>=1.67.1", "opentelemetry-proto>=1.27.0", "grpcio-reflection>=1.62.3", "sqlalchemy>=2.0.36", "psycopg>=3.2.3", "psycopg-binary>=3.2.3", "httpx>=0.27.2", "click>=8.1.7", "lxml>=5.3.0", "types-lxml>=2024.9.16", "opentelemetry-api>=1.27.0", "opentelemetry-exporter-otlp>=1.27.0", "opentelemetry-distro>=0.48b0", "litestar[opentelemetry,standard]>=2.12.1", "faststream[nats,otel]>=0.5.28", "pyyaml>=6.0.2", "nats-py>=2.9.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
dev = ["ruff>=0.7.2"]


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
extend-select = ["I001"]
fixable = ["ALL"]


[tool.pdm.scripts]
alloy = {cmd = "./alloy/alloy-linux-amd64 run alloy/config.alloy --stability.level experimental"}
start-db = {cmd = "docker run --rm --name otel-db -d -p 5432:5432 -e POSTGRES_PASSWORD=postgres postgres:17"}
stop-db = {cmd = "docker stop otel-db"}
start-nats = {cmd = "docker run --rm --name otel-nats -d -p 4222:4222 nats:2.10"}
stop-nats = {cmd = "docker stop otel-nats"}
start-deps = {composite = ["start-db", "start-nats"]}
stop-deps = {composite = ["stop-db", "stop-nats"]}